backend/auth-service/requirements.txt
django>=4.0.0,<5.0.0
djangorestframework>=3.14.0,<4.0.0
djangorestframework-simplejwt>=5.2.0,<6.0.0
django-cors-headers>=3.13.0,<4.0.0
psycopg2-binary>=2.9.3,<3.0.0
gunicorn>=20.1.0,<21.0.0 